Move InvokeMethodListener
